Exit

+++ bot.py
@@ -26,11 +26,12 @@ else:
 
 cluster = MongoClient(DATABASE)
 
-timeout = 600 #seconds
+timeout = 2400 #seconds
 db = cluster["channel-usage"]
-collection = db["channel-usage"]
+channel_usage = db["channel-usage"]
+past_teams = db["past-teams"]
 
-bot = commands.Bot(command_prefix='!')
+bot = commands.Bot(command_prefix='&')
 
 
 def check(author):
@@ -50,7 +51,28 @@ def channels_exist(ctx, vc_list):
 		return 0
 
 
-def remove_players_str(ctx, members, mem_add, players):
+def randomizer(mem_names):
+	random.shuffle(mem_names)
+	mem_str1 = "\n**Team 1**: "
+	mem_str2 = "\n**Team 2**: "
+	count = 1
+	for member in mem_names:
+		if count % 2 == 0:
+			if count == 2:
+				mem_str1 += member + " "
+			else:
+				mem_str1 += ", " + member + " "
+		else:
+			if count == 1:
+				mem_str2 += member + " "
+			else:
+				mem_str2 += ", " + member + " "
+		count+=1
+	response = 'Here are randomized teams. To accept, react with :ok:. To re randomize, react with :arrows_counterclockwise:. To exit, react with :x:.\n'
+	return (response + mem_str1 + mem_str2)
+
+
+def remove_players_str(ctx, members, mem_add):
 	count = 1
 	mem_str = ""
 	for member in members:
@@ -63,20 +85,7 @@ def remove_players_str(ctx, members, mem_add, players):
 		mem_str += str(count) + "  " + member + '\t '
 		count+=1
 
-	response = 'Please select which of the following members should be removed from the game by their numbers. For example, to remove players 4, 8, and 12 from the inhouse game, reply with \'4 8 12\' (without apostrophes)\n\n'
-	return response + mem_str
-
-
-def randomizer(mem_names):
-	random.shuffle(mem_names)
-	mem_str = "Team 1".ljust(25) + "Team 2\n"
-	count = 1
-	for member in mem_names:
-		mem_str += member.ljust(25)
-		if count % 2 == 0:
-			mem_str += '\n'
-		count+=1
-	response = 'Here are randomized teams. To accept, react with the ok reaction. To re randomize, react with the reroll button.\n'
+	response = 'Please select which of the following members should be removed from the game by their numbers. For example, to remove players 4, 8, and 12 from the inhouse game, reply with \'4 8 12\' (without apostrophes).\n\n'
 	return response + mem_str
 
 
@@ -88,19 +97,19 @@ async def inhouse_start(ctx, players: int=10):
 	complete = False
 
 	while time.time() < timeout_start + timeout:
-		resp = f'There are {len(members)+len(mem_add)} selected for the inhouse. If you would like to add more players, react with the plus reaction. If you would like to remove players, react with the minus reaction. If you would like to continue, react with ok reaction.'
+		resp = f'There are {len(members)+len(mem_add)} selected for the inhouse. If you would like to add more players, react with plus sign. If you would like to remove players, react with the minus sign. If you would like to continue, react with :ok:. To quit, react with the :x:.'
 		msg_orig = await ctx.send(resp)
-		reactions = ['\U00002795', '\U00002796', '\U0001f197']
+		reactions = ['\U00002795', '\U00002796', '\U0001f197', '\U0000274c']
 		for reaction in reactions:
 			await msg_orig.add_reaction(emoji=reaction)
 		await asyncio.sleep(0.5)
-		reaction, user = await bot.wait_for('reaction_add', timeout = 120.0)
+		reaction, user = await bot.wait_for('reaction_add', timeout = 600)
 		reac_name = unicodedata.name(reaction.emoji)
 		if reac_name == 'HEAVY MINUS SIGN':
 			await msg_orig.delete()
-			resp = remove_players_str(ctx, members, mem_add, players)
+			resp = remove_players_str(ctx, members, mem_add)
 			msg1 = await ctx.send(resp)
-			msg = await bot.wait_for('message', check=check(ctx.message.author), timeout=120)
+			msg = await bot.wait_for('message', check=check(ctx.message.author), timeout=600)
 			resp_list = list(map(int, msg.content.split()))
 			resp_list.sort(reverse=True)
 			for idx in resp_list:
@@ -114,7 +123,7 @@ async def inhouse_start(ctx, players: int=10):
 			await msg_orig.delete()
 			resp = "Please list the names of the players you would like to add. Since they are currently not connected to a voice channel, they will not be able to be moved to their team channel. For example, to add rwon and Elijah p, reply with \'rwon \"Elijah p\"\' (without apostrophes)\n"
 			msg1 = await ctx.send(resp)
-			msg = await bot.wait_for('message', check=check(ctx.message.author), timeout=120)
+			msg = await bot.wait_for('message', check=check(ctx.message.author), timeout=600)
 			for user in msg.content.split():
 				mem_add.append(user)
 			await msg.delete()
@@ -123,53 +132,66 @@ async def inhouse_start(ctx, players: int=10):
 			complete = True
 			await msg_orig.delete()
 			break
-	if not complete:
-		await ctx.send("Timed out while adding/removing players from the teams. Please try again.")
-	else:
+		elif reac_name == 'CROSS MARK':
+			complete = False
+			await msg_orig.delete()
+			break
+
+	if complete:
 		mem_names = []
 		for member in members:
-			mem_names.append(member.name)
+			if member.nick:
+				mem_names.append(member.nick)
+			else:
+				mem_names.append(member.name)
 		mem_names.extend(mem_add)
 		timeout_start = time.time()
 		while time.time() < timeout_start + timeout:
 			resp = randomizer(mem_names)
 			msg_orig = await ctx.send(resp)
-			reactions = ['\U0001f197', '\U0001f504']
+			reactions = ['\U0001f197', '\U0001f504', '\U0000274c']
 			for reaction in reactions:
 				await msg_orig.add_reaction(reaction)
-			await msg_orig.add_reaction(emoji=reaction)
-			reaction, user = await bot.wait_for('reaction_add', timeout = 60.0)
+			await asyncio.sleep(0.5)
+			reaction, user = await bot.wait_for('reaction_add', timeout = 600.0)
 			reac_name = unicodedata.name(reaction.emoji)
 			if reac_name == 'SQUARED OK':
+				exit = False
 				break
 			elif reac_name == 'ANTICLOCKWISE DOWNWARDS AND UPWARDS OPEN CIRCLE ARROWS':
 				await msg_orig.delete()
 				continue
+			elif reac_name == 'CROSS MARK':
+				await msg_orig.delete()
+				exit = True
+				break
 
-	msg2 = await ctx.send('Teams confirmed! React with OK to move players to each channel.')
-	await msg2.add_reaction(emoji='\U0001f197')
-	await asyncio.sleep(0.5)
+		if not exit:
+			msg2 = await ctx.send('Teams confirmed! React with :ok: to move players to each channel.')
+			await msg2.add_reaction(emoji='\U0001f197')
+			await asyncio.sleep(0.5)
 
-	reaction, user = await bot.wait_for('reaction_add', timeout = 600.0)
+			reaction, user = await bot.wait_for('reaction_add', timeout = 1200.0)
 
-	data = collection.find_one({"_id": ctx.guild.id})
-	team1ch = discord.utils.get(ctx.guild.channels, name=data["team1"])
-	team2ch = discord.utils.get(ctx.guild.channels, name=data["team2"])
+			data = channel_usage.find_one({"_id": ctx.guild.id})
+			team1ch = discord.utils.get(ctx.guild.channels, name=data["team1"])
+			team2ch = discord.utils.get(ctx.guild.channels, name=data["team2"])
 
-	team1 = mem_names[0::2]
-	team2 = mem_names[1::2]
-	for member in members:
-		if member.name in team1:
-			await member.move_to(team1ch)
-		elif member.name in team2:
-			await member.move_to(team2ch)
+			team1 = mem_names[0::2]
+			team2 = mem_names[1::2]
+			for member in members:
+				if member.name in team1 or member.nick in team1:
+					await member.move_to(team1ch)
+				elif member.name in team2 or member.nick in team2:
+					await member.move_to(team2ch)
 
-	await msg2.delete()
-	await ctx.send('Players sent to each team channel! Good luck and have fun!')
+			await msg2.delete()
+			await ctx.send('Players sent to each team channel! Good luck and have fun!')
+
 
 @bot.command(name='endgame', help='Move people back to the lobby channel.')
 async def end_game(ctx):
-	data = collection.find_one({"_id": ctx.guild.id})
+	data = channel_usage.find_one({"_id": ctx.guild.id})
 	lobbych = discord.utils.get(ctx.guild.channels, name=data["lobby"])
 	team1ch = discord.utils.get(ctx.guild.channels, name=data["team1"])
 	team2ch = discord.utils.get(ctx.guild.channels, name=data["team2"])
@@ -178,17 +200,18 @@ async def end_game(ctx):
 	for member in team2ch.members:
 		await member.move_to(lobbych)
 
+
 @bot.command(name='setchannel', help='Set the default voice channels for the lobby and team channels.')
 async def set_channels(ctx, lobby: str, team1: str, team2: str):
 	if channels_exist(ctx, [lobby, team1, team2]):
 		idquery = {"_id": ctx.guild.id}
-		if collection.find(idquery).count():
+		if channel_usage.find(idquery).count():
 			newvalues = { "$set": {"lobby": lobby, "team1": team1, "team2": team2}}
-			collection.update_one(idquery, newvalues)
+			channel_usage.update_one(idquery, newvalues)
 			await ctx.send("Channels were set previously. Updating existing records.")
 		else:
 			post = {"_id": ctx.guild.id, "lobby": lobby, "team1": team1, "team2": team2}
-			collection.insert_one(post)
+			channel_usage.insert_one(post)
 			await ctx.send("Creating default channels for your server.")
 	else:
 		await ctx.send("Invalid channel names provided.")
